Revert "move training data stuff into new directory"
This reverts commit 52c7feb3fc4f2649997d059174ccbb604fdbb1d9.
