+++ src/neural/reader.cc
@@ -25,7 +25,10 @@
   Program grant you additional permission to convey the resulting work.
 */
 
-#include "trainingdata/reader.h"
+#include "neural/reader.h"
+
+#include "neural/encoder.h"
+#include "utils/bititer.h"
 
 namespace lczero {
